// frontend/src/types/weather.ts
export interface HourlyForecastItem {
  id?: string;
  date_str: string;
  time_str: string;
  condition: string;
  icon: string;
  temperature: number;
  humidity: number;
  rain_probability?: number;
  wind_speed?: number;
  wind_direction?: string;
}

export interface DailyForecastItem {
  date_str: string;
  date_short: string;
  day_name: string;
  condition: string;
  icon: string;
  min_temp: number;
  max_temp: number;
  humidity?: number;
  rain_probability?: number;
  wind_speed?: number;
  wind_direction?: string;
  pressure?: number;
  sunrise?: string;
  sunset?: string;
  hourly_breakdown?: HourlyForecastItem[];
}